+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #1244 Added "Body for Sample Invalidation email" field in setup
 - #1231 Add Client ID Column in Batch Listing
 - #1230 Add Client ID Column in Sample Listing
 - #1222 Added User and Security API
@@ -98,6 +99,7 @@ Changelog
 
 **Fixed**
 
+- #1245 Not all clients are shown in clients drop menu for Productivity Reports
 - #1239 Fix and Improve Stickers
 - #1214 Disallow entry of analysis results if the sample is not yet received
 - #1213 Fix instrument notification display in Manage Results View